requirements.txt
Flask==2.3.3
PyMySQL==1.1.0
cryptography==41.0.4
PyJWT==2.8.0
Werkzeug==2.3.7 